Trying onehotting the classification

+++ src/data.py
@@ -109,10 +109,10 @@ class SampleGenerator(object):
                 items.append(int(row.negatives[i]))
                 ratings.append(float(0))  # negative samples get 0 rating
         if self.classification is True:
-            ratings = torch.nn.functional.one_hot(torch.as_tensor(ratings, dtype=torch.int64), num_classes=6)
+            ratings = torch.nn.functional.one_hot(torch.as_tensor(ratings, dtype=torch.float32), num_classes=6)
         dataset = UserItemRatingDataset(user_tensor=torch.LongTensor(users),
                                         item_tensor=torch.LongTensor(items),
-                                        target_tensor=torch.FloatTensor(ratings))
+                                        target_tensor=ratings)
         return DataLoader(dataset, batch_size=batch_size, shuffle=True)
 
     @property
